Rename fixtures.ts to historyAPIFixtures.ts
