+++ ethereum_history_api/oracles/src/historyAPIFixtures.ts
