+++ ethereum_history_api/oracles/src/noir/oracles/transactionOracle/encode.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { TX_OFFSETS, encodeTx } from './encode.js';
-import { loadTxFixture } from '../../../fixtures.js';
+import { loadTxFixture } from '../../../historyAPIFixtures.js';
 import { ETH_TRANSFER_TX_HASH, CHAIN_LINK_TRANSFER_TX_HASH } from '../../../fixtures/historyAPIConfig.js';
 
 // TODO: Add a test for contract creation transaction when we support transactions longer than trie nodes.
